API

LinearWaves.LinearWaves

LinearWaves.dispersion_relation

LinearWaves.group_velocity

LinearWaves.horizontal_displacement

LinearWaves.horizontal_displacement_s

LinearWaves.horizontal_velocity

LinearWaves.horizontal_velocity_s

LinearWaves.phase_velocity

LinearWaves.pressure

LinearWaves.pressure_s

LinearWaves.vertical_displacement

LinearWaves.vertical_displacement_s

LinearWaves.vertical_velocity

LinearWaves.vertical_velocity_s

LinearWaves.wave_profile

LinearWaves.wave_profile_s

LinearWaves.LinearWaves — Module

LinearWaves is an implementation of a linear water wave theory.

LinearWaves.dispersion_relation — Method

dispertion_relation(d, ω; g=LinearWaves.g, ϵ=LinearWaves.tol)

Calculate a wave number k based on depth d, an angular wave frequency ω and gravitational acceleration g for a given accuracy ϵ according to linear wave theory."

LinearWaves.group_velocity — Method

group_velocity(k, ω, d)

Calculate group velocity (celerity) c_g for a regular wave of a number k and an angular frequency ω at depth d.

LinearWaves.horizontal_displacement — Method

horizontal_displacement(x::Real, z::Real, t::Real, a::Real, k::Real, ω::Real, d::Real)

Calculate horizontal particle displacement ξ at position (x, z) and time t for a regular wave of amplitude a, wave number k, and angular wave frequency ω, propagating at constant depth d.

LinearWaves.horizontal_displacement_s — Method

horizontal_displacement_s(x::Real, z::Real, t::Real, a::Real, k::Real, ω::Real, d::Real)

Calculate horizontal particle displacement ξ at position (x, z) and time t for a regular standing wave of amplitude a, wave number k, and angular wave frequency ω, propagating at constant depth d.

LinearWaves.horizontal_velocity — Method

horizontal_velocity(x::Real, z::Real, t::Real, a::Real, k::Real, ω::Real, d:Real)

Calculate horizontal wave velocity u at position x and time t for a regular wave of amplitude a, wave number k, and angular wave frequency ω, propagating at constant depth d.

LinearWaves.horizontal_velocity_s — Method

horizontal_velocity_s(x::Real, z::Real, t::Real, a::Real, k::Real, ω::Real, d:Real)

Calculate horizontal wave velocity u at position x and time t for a regular standing wave of amplitude a, wave number k, and angular wave frequency ω, propagating at constant depth d.

LinearWaves.phase_velocity — Method

phase_velocity(k, ω)

Calculate phase velocity (celerity) c for a regular wave of a number k and an angular frequency ω.

LinearWaves.pressure — Method

pressure(x::Real, z::Real, t::Real, a::Real, k::Real, ω::Real, d::Real)

Calculate pressure p at position (x, z) and time t under a regular wave of amplitude a, wave number k, and angular wave frequency ω, propagating at constant depth d.

LinearWaves.pressure_s — Method

pressure_s(x::Real, z::Real, t::Real, a::Real, k::Real, ω::Real, d::Real)

Calculate pressure p at position (x, z) and time t under a regular standing wave of amplitude a, wave number k, and angular wave frequency ω, propagating at constant depth d.

LinearWaves.vertical_displacement — Method

vertical_displacement(x::Real, z::Real, t::Real, a::Real, k::Real, ω::Real, d::Real)

Calculate vertical particle displacement ζ at position (x, z) and time t for a regular wave of amplitude a, wave number k, and angular wave frequency ω, propagating at constant depth d.

LinearWaves.vertical_displacement_s — Method

vertical_displacement_s(x::Real, z::Real, t::Real, a::Real, k::Real, ω::Real, d::Real)

Calculate vertical particle displacement ζ at position (x, z) and time t for a regular standing wave of amplitude a, wave number k, and angular wave frequency ω, propagating at constant depth d.

LinearWaves.vertical_velocity — Method

vertical_velocity(x::Real, z::Real, t::Real, a::Real, k::Real, ω::Real, d:Real)

Calculate vertical wave velocity v at position (x, z) and time t for a regular wave of amplitude a, wave number k, and angular wave frequency ω, propagating at constant depth d.

LinearWaves.vertical_velocity_s — Method

vertical_velocity_s(x::Real, z::Real, t::Real, a::Real, k::Real, ω::Real, d:Real)

Calculate vertical wave velocity v at position (x, z) and time t for a regular standing wave of amplitude a, wave number k, and angular wave frequency ω, propagating at constant depth d.

LinearWaves.wave_profile — Method

wave_profile(x::Real, t::Real, a::Real, k::Real, ω::Real)

Calculate free-surface elevation η at position x and time t for a regular wave of amplitude a, wave number k, and angular wave frequency ω.

LinearWaves.wave_profile_s — Method

wave_profile_s(x::Real, t::Real, a::Real, k::Real, ω::Real)

Calculate free-surface elevation η at position x and time t for a regular standing wave of amplitude a, wave number k, and angular wave frequency ω.